requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=mkl
brotlipy=0.7.0=py38h7b6447c_1000
bs4=4.9.3=0
ca-certificates=2020.10.14=0
certifi=2020.6.20=py38h06a4308_2
cffi=1.14.3=py38he30daa8_0
chardet=3.0.4=py38_1003
click=7.1.2=py_0
cryptography=3.1.1=py38h1ba5d50_0
cycler=0.10.0=py38_0
dbus=1.13.18=hb2f20db_0
expat=2.2.10=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
glib=2.66.1=h92f7085_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
idna=2.10=py_0
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py38hfd86e86_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1h=h7b6447c_0
pandas=1.1.3=py38he6710b0_0
pcre=8.44=he6710b0_0
pillow=8.0.1=py38he98fc37_0
pip=20.2.4=py38_0
pycparser=2.20=py_2
pymongo=3.11.0=py38he6710b0_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python-dotenv=0.15.0=pyhd8ed1ab_0
python_abi=3.8=1_cp38
pytz=2020.1=py_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
scipy=1.5.2=py38h0b6359f_0
setuptools=50.3.0=py38hb0f4dca_1
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
soupsieve=2.0.1=py_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
tornado=6.0.4=py38h7b6447c_1
urllib3=1.25.11=py_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
argon2-cffi=20.1.0=py38h7b6447c_1
astroid=2.4.2=py38_0
astunparse=1.6.3=pypi_0
async_generator=1.10=py_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=mkl
bleach=3.2.1=py_0
bokeh=2.2.3=py38_0
brotlipy=0.7.0=py38h7b6447c_1000
bs4=4.9.3=0
ca-certificates=2020.10.14=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=py38h06a4308_2
cffi=1.14.3=py38he30daa8_0
chardet=3.0.4=py38_1003
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
colorcet=2.0.2=py_0
cryptography=3.1.1=py38h1ba5d50_0
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h7b6447c_0
dask=2.30.0=py_0
dask-core=2.30.0=py_0
datashader=0.11.1=py_0
datashape=0.5.4=py38_1
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
distributed=2.30.0=py38_0
entrypoints=0.3=py38_0
expat=2.2.10=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.8.3=py_0
gast=0.3.3=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.33.2=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=pypi_0
heapdict=1.0.1=py_0
holoviews=1.13.5=pyhd3eb1b0_0
icu=58.2=he6710b0_3
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_1
isort=5.6.4=py_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38_0
jupyterlab_pygments=0.1.2=py_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=py38hfd86e86_0
lazy-object-proxy=1.4.3=py38h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libllvm10=10.0.1=he513fc3_3
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.34.0=py38h4f45e52_2
locket=0.2.0=py38_1
lz4-c=1.9.2=heb0550a_3
markdown=3.3.2=py38_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h817c723_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
msgpack-python=1.0.0=py38hfd86e86_1
multipledispatch=0.6.0=py38_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
notebook=6.1.4=py38_0
numba=0.51.2=py38hc5bc63f_0
numpy=1.18.5=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1h=h7b6447c_0
opt-einsum=3.3.0=pypi_0
packaging=20.4=py_0
pandas=1.1.3=py38he6710b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py38_1
panel=0.9.7=py_0
param=1.9.3=py_0
parso=0.7.0=py_0
partd=1.1.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=8.0.1=py38he98fc37_0
pip=20.2.4=py38_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.13.0=pypi_0
psutil=5.7.2=py38h7b6447c_0
ptyprocess=0.6.0=py38_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pyct=0.4.8=py38_0
pygments=2.7.1=py_0
pylint=2.6.0=py38_0
pymongo=3.11.0=py38he6710b0_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python-dotenv=0.15.0=pyhd8ed1ab_0
python_abi=3.8=1_cp38
pytz=2020.1=py_0
pyviz_comms=0.7.6=py_0
pyyaml=5.3.1=py38h7b6447c_1
pyzmq=19.0.2=py38he6710b0_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
seaborn=0.11.0=py_0
send2trash=1.5.0=py38_0
setuptools=50.3.0=py38hb0f4dca_1
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
sortedcontainers=2.2.2=py_0
soupsieve=2.0.1=py_0
sqlite=3.33.0=h62c20be_0
tbb=2020.2=hc9558a2_0
tblib=1.7.0=py_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow=2.3.1=pypi_0
tensorflow-estimator=2.3.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
toolz=0.11.1=py_0
tornado=6.0.4=py38h7b6447c_1
tqdm=4.50.2=py_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
umap-learn=0.4.6=py38h32f6830_0
urllib3=1.25.11=py_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py38_0
wrapt=1.11.2=py38h7b6447c_0
xarray=0.16.1=py_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zict=2.0.0=py_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
